+++ WHATSNEW.md
@@ -1,5 +1,6 @@
 ### Added
 
+- Support for Hashicorp Vault service to store sensitive inventory data
 - Add license, repository and classifiers in pyproject.toml
 
 ### Changed